Selection Controller, Outline and recticle
Added a selection controller so we can always tell if the player has something selected or not and what that something is.
Also now the player  can tell if they can interact with something or not because it will highlight if they can.
Finally, hid the mouse when locked and added pretty icons for the new recticle that change depending on state.